moved saving file into utils, got rid of repetitive code, better handled autosave fails
