+++ clients/cli/src/edit.rs
@@ -11,14 +11,13 @@ use lockbook_core::{
 };
 
 use crate::utils::{
-    edit_file_with_editor, exit_with, get_account_or_exit, get_config, set_up_auto_save,
-    stop_auto_save,
+    edit_file_with_editor, exit_with, get_account_or_exit, get_config, save_temp_file_contents,
+    set_up_auto_save, stop_auto_save,
 };
 use crate::{
     COULD_NOT_DELETE_OS_FILE, COULD_NOT_READ_OS_FILE, COULD_NOT_WRITE_TO_OS_FILE,
     DOCUMENT_TREATED_AS_FOLDER, FILE_NOT_FOUND, SUCCESS, UNEXPECTED_ERROR,
 };
-use lockbook_core::model::file_metadata::FileMetadata;
 
 pub fn edit(file_name: &str) {
     get_account_or_exit();
@@ -93,31 +92,7 @@ pub fn edit(file_name: &str) {
     stop_auto_save(watcher, file_location.clone());
 
     if edit_was_successful {
-        let secret = fs::read_to_string(temp_file_path).unwrap_or_else(|_| {
-            exit_with(
-                &format!(
-                    "Failed to read from temporary file, check {}",
-                    file_location
-                ),
-                COULD_NOT_READ_OS_FILE,
-            )
-        });
-
-        match write_document(&get_config(), file_metadata.id, secret.as_bytes()) {
-            Ok(_) => exit_with(
-                "Document encrypted and saved. Cleaning up temporary file.",
-                SUCCESS,
-            ),
-            Err(err) => match err {
-                CoreError::UiError(WriteToDocumentError::NoAccount)
-                | CoreError::UiError(WriteToDocumentError::FileDoesNotExist)
-                | CoreError::UiError(WriteToDocumentError::FolderTreatedAsDocument)
-                | CoreError::Unexpected(_) => exit_with(
-                    &format!("Unexpected error saving file: {:#?}", err),
-                    UNEXPECTED_ERROR,
-                ),
-            },
-        }
+        save_temp_file_contents(file_metadata, &file_location, temp_file_path, false)
     } else {
         eprintln!("Your editor indicated a problem, aborting and cleaning up");
     }
@@ -128,58 +103,4 @@ pub fn edit(file_name: &str) {
             COULD_NOT_DELETE_OS_FILE,
         )
     });
-}
-
-pub fn save_file_to_core(
-    file_metadata: FileMetadata,
-    file_location: &String,
-    temp_file_path: &Path,
-    silent: bool,
-) {
-    let secret = match fs::read_to_string(temp_file_path) {
-        Ok(content) => content.into_bytes(),
-        Err(err) => {
-            if !silent {
-                exit_with(
-                    &format!(
-                        "Could not read from temporary file, not deleting {}, err: {:#?}",
-                        file_location, err
-                    ),
-                    UNEXPECTED_ERROR,
-                )
-            } else {
-                return;
-            }
-        }
-    };
-
-    match write_document(&get_config(), file_metadata.id, &secret) {
-        Ok(_) => {
-            if !silent {
-                exit_with(
-                    "Document encrypted and saved. Cleaning up temporary file.",
-                    SUCCESS,
-                )
-            } else {
-            }
-        }
-        Err(err) => {
-            if !silent {
-                match err {
-                    CoreError::Unexpected(msg) => exit_with(&msg, UNEXPECTED_ERROR),
-                    CoreError::UiError(WriteToDocumentError::NoAccount) => exit_with(
-                        "Unexpected: No account! Run init or import to get started!",
-                        UNEXPECTED_ERROR,
-                    ),
-                    CoreError::UiError(WriteToDocumentError::FileDoesNotExist) => {
-                        exit_with("Unexpected: FileDoesNotExist", UNEXPECTED_ERROR)
-                    }
-                    CoreError::UiError(WriteToDocumentError::FolderTreatedAsDocument) => {
-                        exit_with("Unexpected: CannotWriteToFolder", UNEXPECTED_ERROR)
-                    }
-                }
-            } else {
-            }
-        }
-    }
 }
